+++ package.json
@@ -10,13 +10,13 @@
     "url": "https://github.com/joshraphael/vscode-rascript"
   },
   "engines": {
-    "vscode": "^1.97.0"
+    "vscode": "^1.102.0"
   },
   "categories": [
     "Programming Languages"
   ],
   "activationEvents": [],
-  "main": "./out/extension/extension.js",
+  "main": "./out/node/extension.js",
   "browser": "./out/web/extension.js",
   "contributes": {
     "configuration": {
@@ -52,8 +52,8 @@
   },
   "scripts": {
     "vscode:prepublish": "npm run compile",
-    "compile": "tsc -p ./ && webpack --mode production --devtool hidden-source-map",
-    "watch": "tsc -watch -p ./",
+    "compile": "tsc -p ./src/shared && tsc -p ./src/node && webpack",
+    "watch": "tsc -watch -p ./src/node",
     "pretest": "npm run compile && npm run lint:check && npm run format:check",
     "lint": "eslint --ext .js,.ts,.tsx src --max-warnings=0 --fix",
     "format": "prettier --write 'src/**/*.{json,md,js,ts,tsx}'",
@@ -63,23 +63,23 @@
     "browser": "npm run pretest && vscode-test-web --extensionDevelopmentPath=. ."
   },
   "devDependencies": {
-    "@types/vscode": "^1.97.0",
+    "@types/vscode": "^1.102.0",
     "@types/mocha": "^10.0.10",
     "@types/node": "20.x",
-    "@typescript-eslint/eslint-plugin": "^8.34.0",
-    "@typescript-eslint/parser": "^8.34.0",
+    "@typescript-eslint/eslint-plugin": "^8.38.0",
+    "@typescript-eslint/parser": "^8.38.0",
     "eslint": "^9.19.0",
     "eslint-config-prettier": "^8.6.0",
-    "typescript": "^5.7.3",
+    "typescript": "^5.8.2",
     "@vscode/test-cli": "^0.0.11",
     "@vscode/test-electron": "^2.4.1",
-    "@vscode/test-web": "^0.0.70",
-    "webpack": "^5.75.0",
-    "webpack-cli": "^5.0.1",
-    "ts-loader": "^7.0.5",
+    "@vscode/test-web": "^0.0.71",
+    "webpack": "^5.101.0",
+    "webpack-cli": "^6.0.1",
+    "ts-loader": "^9.5.2",
     "prettier": "2.8.3"
   },
   "dependencies": {
-    "vscode-languageclient": "^9.0.1"
+    "vscode-languageclient": "^10.0.0-next.16"
   }
 }
