+++ src/shared/extension.ts
@@ -1,83 +1,9 @@
 import * as vscode from "vscode";
-
-import {
-  LanguageClient,
-  LanguageClientOptions,
-  ServerOptions,
-} from "vscode-languageclient/node";
 import { builtinFunctionDefinitions } from "./functionDefinitions";
 import * as parser from "./parser";
 import * as models from "./models";
 
-let client: LanguageClient;
-
-export function activate(context: vscode.ExtensionContext) {
-  const rascriptLanguageServer =
-    vscode.workspace.getConfiguration("rascript").languageServer;
-  setup(context, rascriptLanguageServer);
-}
-
-async function setup(
-  context: vscode.ExtensionContext,
-  rascriptLanguageServer: string
-) {
-  const fileUri = vscode.Uri.file(rascriptLanguageServer);
-  if (rascriptLanguageServer === undefined || rascriptLanguageServer === "") {
-    vscode.window.showInformationMessage("No Language Server specified");
-    localExtension(context);
-    return;
-  }
-  languageServer(context, rascriptLanguageServer);
-}
-
-function languageServer(
-  context: vscode.ExtensionContext,
-  rascriptLanguageServer: string
-) {
-  let serverOptions: ServerOptions = {
-    run: { command: rascriptLanguageServer },
-    debug: { command: rascriptLanguageServer },
-  };
-
-  // Options to control the language client
-  let clientOptions: LanguageClientOptions = {
-    // Register the server for plain text documents
-    documentSelector: [
-      {
-        pattern: "**/*.rascript",
-      },
-    ],
-    synchronize: {
-      configurationSection: "rascriptLanguageServer",
-      fileEvents: vscode.workspace.createFileSystemWatcher("**/*.rascript"),
-    },
-  };
-
-  // Create the language client and start the client.
-  client = new LanguageClient(
-    "rascript-language-server",
-    "RAScript Language Server",
-    serverOptions,
-    clientOptions
-  );
-
-  // Start the client. This will also launch the server
-  client
-    .start()
-    .then(() => {
-      vscode.window.showInformationMessage(
-        "Language Server started from: " + rascriptLanguageServer
-      );
-    })
-    .catch((error) => {
-      vscode.window.showInformationMessage(
-        "Failed to start language server: " + error
-      );
-      localExtension(context);
-    });
-}
-
-function localExtension(context: vscode.ExtensionContext) {
+export function localExtension(context: vscode.ExtensionContext) {
   const definitions = vscode.languages.registerDefinitionProvider("rascript", {
     provideDefinition(document, position, token) {
       let text = document.getText();
@@ -377,11 +303,4 @@ function localExtension(context: vscode.ExtensionContext) {
   );
 
   context.subscriptions.push(autocomplete, hover, definitions);
-}
-
-export function deactivate(): Thenable<void> | undefined {
-  if (!client) {
-    return undefined;
-  }
-  return client.stop();
 }
