+++ src/shared/functionDefinitions.ts
